refactor(step1): LineSectionAcceptanceTest 리팩터링
- 지하철 구간을 관리하는 테스트로 통합
- static 메서드 분리
